requirements.txt

# Core ML/CV Libraries
torch>=1.13.0
torchvision>=0.14.0
numpy>=1.21.0
Pillow>=9.0.0
opencv-python>=4.6.0

# Google Drive API
google-api-python-client>=2.70.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.8.0
google-auth>=2.15.0

# AWS S3 (Optional)
boto3>=1.26.0
botocore>=1.29.0

# Data Processing
tqdm>=4.64.0
requests>=2.28.0
pyyaml>=6.0
h5py>=3.7.0
pathlib2>=2.3.0
humanfriendly>=9.2.0

# System Monitoring
psutil>=5.9.0

# Image Processing & Augmentation
albumentations>=1.3.0
scikit-image>=0.19.0

# Utilities
pathlib2>=2.3.7
python-dotenv>=0.21.0
rich>=12.6.0

# Testing & Development
pytest>=7.2.0
pytest-cov>=4.0.0
black>=22.10.0
flake8>=5.0.0

# Configuration and logging
pyyaml>=5.4.0
loguru>=0.5.0

# Optional: For advanced features
tensorboard>=2.11.0
wandb>=0.13.0
matplotlib>=3.6.0
seaborn>=0.12.0
